update spring data to 1.2.6 & ebean 11.6.1

+++ src/main/java/org/springframework/data/ebean/querychannel/EbeanQueryChannelService.java
@@ -154,7 +154,7 @@ public class EbeanQueryChannelService {
     Assert.notNull(expressionList, "expressionList must not null");
     Assert.notNull(pageable, "pageable must not null");
     return expressionList.setMaxRows(pageable.getPageSize())
-        .setFirstRow(pageable.getOffset())
+        .setFirstRow((int) pageable.getOffset())
         .setOrder(Converters.convertToEbeanOrderBy(pageable.getSort()));
   }
